+++ seantis/people/browser/list.py
@@ -9,7 +9,7 @@ from seantis.people.interfaces import IList
 from seantis.people.browser import BaseView, Renderer
 from seantis.people.content.list import ListFilter, LetterFilter
 from seantis.people.supermodel import (
-    get_schema_columns
+    get_schema_columns, compound_columns
 )
 
 
@@ -25,7 +25,7 @@ class ListView(BaseView):
 
     def update(self):
         if self.has_people:
-            self.renderer = Renderer(self.schema)
+            self.renderer = Renderer(self.schema, redirects=compound_columns)
 
     @property
     def has_people(self):